adding data transformation pipeline and flow. And fixing Logs configuration

+++ src/email_classification/components/data_ingestion_component.py
@@ -5,7 +5,6 @@ import gdown
 
 from src.email_classification.entity.config_entity import DataIngestionConfig
 from src.email_classification.constants import constants
-from src.email_classification.utils.common import get_size
 from src.email_classification.logger import logging
 from src.email_classification.exception import CustomException
 
@@ -24,17 +23,18 @@ class DataIngestion:
         try: 
             dataset_url = self.config.source_URL
             zip_download_dir = self.config.local_data_file
-            os.makedirs(self.config.unzip_dir, exist_ok=True)
-            logger.info(f"Downloading data from {dataset_url} into file {zip_download_dir}")
+            if not os.path.exists(self.config.unzip_dir):
+                os.makedirs(self.config.unzip_dir, exist_ok=True)
+                logger.info(f"Downloading data from {dataset_url} into file {zip_download_dir}")
 
-            file_id = dataset_url.split("/")[-2]
-            prefix = constants.GOOGLE_DRIVE_DOWNLOAD_PREFIX_URL
-            gdown.download(prefix+file_id,zip_download_dir)
+                file_id = dataset_url.split("/")[-2]
+                prefix = constants.GOOGLE_DRIVE_DOWNLOAD_PREFIX_URL
+                gdown.download(prefix+file_id,zip_download_dir)
 
-            logger.info(f"Downloaded data from {dataset_url} into file {zip_download_dir}")
+                logger.info(f"Downloaded data from {dataset_url} into file {zip_download_dir}")
 
         except Exception as e:
-            raise e
+            raise CustomException(e, sys)
         
     
 
@@ -46,7 +46,8 @@ class DataIngestion:
         """
         unzip_path = self.config.unzip_dir
         os.makedirs(unzip_path, exist_ok=True)
-        with zipfile.ZipFile(self.config.local_data_file, 'r') as zip_ref:
-            zip_ref.extractall(unzip_path)
-        os.remove(self.config.local_data_file)
+        if os.path.exists(self.config.local_data_file):
+            with zipfile.ZipFile(self.config.local_data_file, 'r') as zip_ref:
+                zip_ref.extractall(unzip_path)
+            os.remove(self.config.local_data_file)
